pyproject.toml
[project]
name = "apt-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogoogle>=5.17.0",
    "aiogram>=3.22.0",
    "aioschedule>=0.5.2",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "google-auth-oauthlib>=1.2.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "pyjwt>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "sqlalchemy-stubs>=0.4",
]


[tool.uv]
cache-dir = ".cache"
